feat : Editor 폴더 생성, 랜더링 시작-id 가져오기, 더미데이터 추가
